+++ main_parameters.c
@@ -7,8 +7,8 @@
 
 int main(int argc, char* argv[]) {
 
-    int fc = 1;
-    sscanf(argv[3], "%i", &fc);
+    double fc = 1.0;
+    sscanf(argv[3], "%lf", &fc);
 	
     params P = motion_params(argv[1], argv[2], fc);	
 	printf("%le %le %le %le %le %le", P.gamma, P.alpha, P.beta, P.A, P.B, P.C);
